requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
python-multipart==0.0.9
PyMuPDF==1.23.25
python-pptx==0.6.23
langchain==0.1.12
langchain-community==0.0.28
chromadb==0.4.24
sentence-transformers==2.5.1
pydantic-settings==2.2.1
langchain-openai==0.0.8
transformers>=4.38.0
torch>=2.0.0
accelerate>=0.27.0
